Refactor icon organization
